backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-multipart
openai
python-dotenv
